Isolate trace allocation tracker implementation
